refactor(cli): migrate argument parsing to clap
Replace the handrolled `parse_cli_args_from` loop with a `clap`-derived
parser. Conflict checks (`--file` vs `--path`/`-r`/`-w`/`--all-files`,
`--all-files` vs `--path`/`-r`/`-w`) are now declarative; unknown flags
are no longer silently ignored. The public `CliArgs` POJO is unchanged
so the rest of the binary needs no updates.
