+++ Cargo.toml
@@ -30,6 +30,9 @@ chrono = { version = "0.4", default-features = false, features = ["std", "clock"
 thiserror = "2.0"
 anyhow = "1.0"
 
+# CLI parsing
+clap = { version = "4", features = ["derive"] }
+
 # Utilities
 directories = "6.0"
 ureq = { version = "3", features = ["json"] }
