+++ src/theme/mod.rs
@@ -7,7 +7,6 @@ use std::{process::Command, sync::OnceLock};
 use ratatui::style::Color;
 use two_face::theme::EmbeddedThemeName;
 
-use crate::config::config_path_hint;
 use crate::syntax::SyntaxHighlighter;
 
 /// Complete color theme for the application
@@ -1573,35 +1572,6 @@ const THEME_CHOICES: [(&str, ThemeArg); 23] = [
     ("everforest-light", ThemeArg::EverforestLight),
 ];
 
-/// CLI arguments parsed from command line
-#[derive(Debug, Clone, Default)]
-pub struct CliArgs {
-    pub theme: Option<ThemeArg>,
-    pub appearance: Option<AppearanceArg>,
-    /// Output to stdout instead of clipboard when exporting
-    pub output_to_stdout: bool,
-    /// Skip checking for updates on startup
-    pub no_update_check: bool,
-    /// Commit/revision range to review
-    pub revisions: Option<String>,
-    /// Skip commit selector and review uncommitted changes directly
-    pub working_tree: bool,
-    /// Filter diff to a specific file or directory path
-    pub path_filter: Option<String>,
-    /// Open a single file or directory for annotation (no VCS required)
-    pub file_path: Option<String>,
-    /// Whole-repo annotation mode (`--all-files` / `-A`). Lists every tracked
-    /// file via `git ls-files` and renders them in context-only mode.
-    pub all_files: bool,
-    /// Direct pull request target from `tuicr pr <target>`.
-    pub pr_target: Option<String>,
-    /// Override the GitHub repository used for all PR operations. Useful when
-    /// the local `origin` points at a fork (e.g. `agavra/slatedb`) but PRs
-    /// live on the upstream (`slatedb/slatedb`). Accepts any URL form
-    /// `parse_github_remote_url` understands (HTTPS, SCP-SSH, ssh://).
-    pub repo_url: Option<String>,
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum AppearanceArg {
     Light,
@@ -1631,7 +1601,7 @@ impl ThemeArg {
         })
     }
 
-    fn valid_values_display() -> String {
+    pub(crate) fn valid_values_display() -> String {
         Self::choices()
             .iter()
             .map(|(name, _)| *name)
@@ -1656,7 +1626,7 @@ impl AppearanceArg {
         })
     }
 
-    fn valid_values_display() -> String {
+    pub(crate) fn valid_values_display() -> String {
         Self::choices()
             .iter()
             .map(|(name, _)| *name)
@@ -1949,379 +1919,10 @@ fn shift_lightness(c: Color, amount: i32) -> Color {
     }
 }
 
-/// Print version and exit
-fn print_version() -> ! {
-    println!("tuicr {}", env!("CARGO_PKG_VERSION"));
-    std::process::exit(0);
-}
-
-/// Print help message and exit
-fn print_help() -> ! {
-    let name = std::env::args()
-        .next()
-        .and_then(|p| {
-            std::path::Path::new(&p)
-                .file_name()
-                .map(|s| s.to_string_lossy().into_owned())
-        })
-        .unwrap_or_else(|| "tuicr".to_string());
-    let valid_values = ThemeArg::valid_values_display();
-    let appearance_values = AppearanceArg::valid_values_display();
-    let config_path = config_path_hint();
-    println!(
-        "tuicr - A code review TUI with vim keybindings. Export to GitHub or clipboard.
-
-Usage: {name} [OPTIONS]
-
-Options:
-  -r, --revisions <REVSET>  Commit range/Revset to review (syntax depends on VCS backend)
-  --theme <THEME>        Color theme to use
-                          Valid values: {valid_values}
-  --appearance <MODE>    Appearance mode for default theme
-                         Valid values: {appearance_values}
-                         Used when no explicit theme is set
-                         Precedence: --appearance > {config_path} > system
-  -p, --path <PATH>     Filter diff to a specific file or directory
-  -w, --working-tree     Include uncommitted changes (skip commit selector when used alone,
-                         combine with commits when used with -r)
-  --file <PATH>          Open a file or directory for annotation (no VCS required)
-  -A, --all-files        Review every tracked file in the cwd's git repo
-  --stdout               Output to stdout instead of clipboard when exporting
-  --no-update-check      Skip checking for updates on startup
-  --repo-url <URL>       Override the GitHub repo used for PR operations
-                         (e.g. when origin is a fork and PRs live on upstream).
-                         Accepts HTTPS, SCP-style SSH, or ssh:// URLs.
-  -V, --version          Print version
-  -h, --help             Print this help message
-
-Press ? in the application for keybinding help."
-    );
-    std::process::exit(0);
-}
-
-/// Parse CLI arguments from command line
-///
-/// We use a handrolled argument parser instead of clap to keep binary size
-/// small and build times fast. If we end up needing more complex argument
-/// handling, we can revisit this decision.
-pub fn parse_cli_args() -> CliArgs {
-    let args: Vec<String> = std::env::args().collect();
-    parse_cli_args_from(&args).unwrap_or_else(|err| {
-        eprintln!("Error: {err}");
-        std::process::exit(2);
-    })
-}
-
-/// Reject `--repo-url` values that don't parse as a GitHub remote URL so the
-/// failure is surfaced at startup rather than when the PR tab is opened.
-fn validate_repo_url(value: &str) -> Result<(), String> {
-    if crate::forge::github::gh::parse_github_remote_url(value).is_some() {
-        return Ok(());
-    }
-    Err(format!(
-        "--repo-url value '{value}' is not a recognized GitHub URL. \
-         Expected forms: https://github.com/owner/repo, git@github.com:owner/repo, \
-         or ssh://git@github.com/owner/repo"
-    ))
-}
-
-fn parse_cli_args_from(args: &[String]) -> Result<CliArgs, String> {
-    let mut cli_args = CliArgs::default();
-
-    // Subcommand form: `tuicr pr <target>`. We special-case it before the
-    // flag loop because `pr` is a positional token, not a flag. Only the
-    // first positional position counts; subsequent arguments after the
-    // target are treated as ordinary flags (theme, appearance, etc).
-    if args.len() >= 2 && args[1] == "pr" {
-        let target = args.get(2).ok_or_else(|| {
-            "tuicr pr requires a target: <number>, <owner/repo#N>, or a PR URL".to_string()
-        })?;
-        if target.starts_with('-') {
-            return Err(
-                "tuicr pr requires a target: <number>, <owner/repo#N>, or a PR URL".to_string(),
-            );
-        }
-        cli_args.pr_target = Some(target.clone());
-    }
-
-    for i in 0..args.len() {
-        // Handle --version / -V
-        if args[i] == "--version" || args[i] == "-V" {
-            print_version();
-        }
-
-        // Handle --help / -h
-        if args[i] == "--help" || args[i] == "-h" {
-            print_help();
-        }
-
-        // Handle --stdout
-        if args[i] == "--stdout" {
-            cli_args.output_to_stdout = true;
-        }
-
-        // Handle --no-update-check
-        if args[i] == "--no-update-check" {
-            cli_args.no_update_check = true;
-        }
-
-        // Handle -w / --working-tree
-        if args[i] == "-w" || args[i] == "--working-tree" {
-            cli_args.working_tree = true;
-        }
-
-        // Handle --theme value
-        if args[i] == "--theme" {
-            let valid_values = ThemeArg::valid_values_display();
-            let value = args
-                .get(i + 1)
-                .ok_or_else(|| format!("--theme requires a value ({valid_values})"))?;
-
-            if value.starts_with('-') {
-                return Err(format!("--theme requires a value ({valid_values})"));
-            }
-
-            cli_args.theme = ThemeArg::from_str(value)
-                .ok_or_else(|| format!("Unknown theme '{value}'. Valid options: {valid_values}"))
-                .map(Some)?;
-        }
-        // Handle --theme=value
-        if let Some(value) = args[i].strip_prefix("--theme=") {
-            let valid_values = ThemeArg::valid_values_display();
-            if value.is_empty() {
-                return Err(format!("--theme requires a value ({valid_values})"));
-            }
-
-            cli_args.theme = ThemeArg::from_str(value)
-                .ok_or_else(|| format!("Unknown theme '{value}'. Valid options: {valid_values}"))
-                .map(Some)?;
-        }
-
-        // Handle --appearance value
-        if args[i] == "--appearance" {
-            let valid_values = AppearanceArg::valid_values_display();
-            let value = args
-                .get(i + 1)
-                .ok_or_else(|| format!("--appearance requires a value ({valid_values})"))?;
-
-            if value.starts_with('-') {
-                return Err(format!("--appearance requires a value ({valid_values})"));
-            }
-
-            cli_args.appearance = AppearanceArg::from_str(value)
-                .ok_or_else(|| {
-                    format!("Unknown appearance '{value}'. Valid options: {valid_values}")
-                })
-                .map(Some)?;
-        }
-
-        // Handle --appearance=value
-        if let Some(value) = args[i].strip_prefix("--appearance=") {
-            let valid_values = AppearanceArg::valid_values_display();
-            if value.is_empty() {
-                return Err(format!("--appearance requires a value ({valid_values})"));
-            }
-
-            cli_args.appearance = AppearanceArg::from_str(value)
-                .ok_or_else(|| {
-                    format!("Unknown appearance '{value}'. Valid options: {valid_values}")
-                })
-                .map(Some)?;
-        }
-
-        // Handle -p / --path value
-        if args[i] == "-p" || args[i] == "--path" {
-            let value = args
-                .get(i + 1)
-                .ok_or_else(|| "--path requires a file or directory path".to_string())?;
-            if value.starts_with('-') {
-                return Err("--path requires a file or directory path".to_string());
-            }
-            cli_args.path_filter = Some(value.clone());
-        }
-        // Handle --path=value
-        if let Some(value) = args[i].strip_prefix("--path=") {
-            if value.is_empty() {
-                return Err("--path requires a file or directory path".to_string());
-            }
-            cli_args.path_filter = Some(value.to_string());
-        }
-
-        // Handle --file value
-        if args[i] == "--file" {
-            let value = args
-                .get(i + 1)
-                .ok_or_else(|| "--file requires a file path".to_string())?;
-            if value.starts_with('-') {
-                return Err("--file requires a file path".to_string());
-            }
-            cli_args.file_path = Some(value.clone());
-        }
-        // Handle --file=value
-        if let Some(value) = args[i].strip_prefix("--file=") {
-            if value.is_empty() {
-                return Err("--file requires a file path".to_string());
-            }
-            cli_args.file_path = Some(value.to_string());
-        }
-
-        // Handle --all-files / -A
-        if args[i] == "--all-files" || args[i] == "-A" {
-            cli_args.all_files = true;
-        }
-
-        // Handle --repo-url value
-        if args[i] == "--repo-url" {
-            let value = args
-                .get(i + 1)
-                .ok_or_else(|| "--repo-url requires a value (e.g. https://github.com/owner/repo or git@github.com:owner/repo)".to_string())?;
-            if value.starts_with('-') {
-                return Err("--repo-url requires a value (e.g. https://github.com/owner/repo or git@github.com:owner/repo)".to_string());
-            }
-            validate_repo_url(value)?;
-            cli_args.repo_url = Some(value.clone());
-        }
-        // Handle --repo-url=value
-        if let Some(value) = args[i].strip_prefix("--repo-url=") {
-            if value.is_empty() {
-                return Err("--repo-url requires a value (e.g. https://github.com/owner/repo or git@github.com:owner/repo)".to_string());
-            }
-            validate_repo_url(value)?;
-            cli_args.repo_url = Some(value.to_string());
-        }
-
-        // Handle -r / --revisions value
-        if args[i] == "-r" || args[i] == "--revisions" {
-            if let Some(value) = args.get(i + 1) {
-                cli_args.revisions = Some(value.clone());
-            } else {
-                eprintln!("Warning: {0} requires a value", args[i]);
-            }
-        }
-        // Handle --revisions=value
-        if let Some(value) = args[i].strip_prefix("--revisions=") {
-            cli_args.revisions = Some(value.to_string());
-        }
-    }
-
-    Ok(cli_args)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use std::collections::HashSet;
-
-    fn parse_for_test(args: &[&str]) -> Result<CliArgs, String> {
-        let args = args.iter().map(|s| s.to_string()).collect::<Vec<_>>();
-        parse_cli_args_from(&args)
-    }
-
-    #[test]
-    fn should_parse_theme_when_provided() {
-        let parsed = parse_for_test(&["tuicr", "--theme", "light"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::Light));
-    }
-
-    #[test]
-    fn should_parse_catppuccin_themes() {
-        let parsed = parse_for_test(&["tuicr", "--theme", "catppuccin-mocha"])
-            .expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::CatppuccinMocha));
-
-        let parsed =
-            parse_for_test(&["tuicr", "--theme=catppuccin-latte"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::CatppuccinLatte));
-    }
-
-    #[test]
-    fn should_parse_ayu_light_theme() {
-        let parsed =
-            parse_for_test(&["tuicr", "--theme", "ayu-light"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::AyuLight));
-    }
-
-    #[test]
-    fn should_parse_onedark_theme() {
-        let parsed =
-            parse_for_test(&["tuicr", "--theme", "onedark"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::Onedark));
-    }
-
-    #[test]
-    fn should_parse_gruvbox_themes() {
-        let parsed =
-            parse_for_test(&["tuicr", "--theme", "gruvbox-dark"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::GruvboxDark));
-
-        let parsed =
-            parse_for_test(&["tuicr", "--theme=gruvbox-light"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::GruvboxLight));
-    }
-
-    #[test]
-    fn should_leave_theme_none_when_not_provided() {
-        let parsed = parse_for_test(&["tuicr"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, None);
-    }
-
-    #[test]
-    fn should_parse_working_tree_short_flag() {
-        let parsed = parse_for_test(&["tuicr", "-w"]).expect("parse should succeed");
-        assert!(parsed.working_tree);
-    }
-
-    #[test]
-    fn should_parse_working_tree_long_flag() {
-        let parsed = parse_for_test(&["tuicr", "--working-tree"]).expect("parse should succeed");
-        assert!(parsed.working_tree);
-    }
-
-    #[test]
-    fn should_default_working_tree_to_false() {
-        let parsed = parse_for_test(&["tuicr"]).expect("parse should succeed");
-        assert!(!parsed.working_tree);
-    }
-
-    #[test]
-    fn should_parse_working_tree_with_revisions() {
-        let parsed =
-            parse_for_test(&["tuicr", "-w", "-r", "HEAD~3..HEAD"]).expect("parse should succeed");
-        assert!(parsed.working_tree);
-        assert_eq!(parsed.revisions, Some("HEAD~3..HEAD".to_string()));
-    }
-
-    #[test]
-    fn should_error_for_invalid_theme_in_separate_arg() {
-        let err = parse_for_test(&["tuicr", "--theme", "nope"]).expect_err("parse should fail");
-        assert!(err.contains("Unknown theme 'nope'"));
-    }
-
-    #[test]
-    fn should_error_for_invalid_theme_in_equals_arg() {
-        let err = parse_for_test(&["tuicr", "--theme=nope"]).expect_err("parse should fail");
-        assert!(err.contains("Unknown theme 'nope'"));
-    }
-
-    #[test]
-    fn should_error_when_theme_value_missing() {
-        let err = parse_for_test(&["tuicr", "--theme"]).expect_err("parse should fail");
-        assert!(err.contains("--theme requires a value"));
-    }
-
-    #[test]
-    fn should_parse_appearance_when_provided() {
-        let parsed =
-            parse_for_test(&["tuicr", "--appearance", "system"]).expect("parse should succeed");
-        assert_eq!(parsed.appearance, Some(AppearanceArg::System));
-    }
-
-    #[test]
-    fn should_error_for_invalid_appearance() {
-        let err =
-            parse_for_test(&["tuicr", "--appearance", "nope"]).expect_err("parse should fail");
-        assert!(err.contains("Unknown appearance 'nope'"));
-    }
 
     #[test]
     fn should_roundtrip_all_canonical_theme_values() {
@@ -2484,17 +2085,6 @@ mod tests {
     }
 
     #[test]
-    fn should_parse_everforest_themes() {
-        let parsed =
-            parse_for_test(&["tuicr", "--theme", "everforest-dark"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::EverforestDark));
-
-        let parsed =
-            parse_for_test(&["tuicr", "--theme=everforest-light"]).expect("parse should succeed");
-        assert_eq!(parsed.theme, Some(ThemeArg::EverforestLight));
-    }
-
-    #[test]
     fn should_resolve_everforest_dark_to_gruvbox_dark_syntect_theme() {
         let theme = resolve_theme(ThemeArg::EverforestDark);
         assert_eq!(theme.syntect_theme, EmbeddedThemeName::GruvboxDark);
@@ -2582,206 +2172,5 @@ mod tests {
     fn should_return_accent_for_non_rgb_blend_inputs() {
         let accent = Color::Rgb(100, 110, 120);
         assert_eq!(blend(Color::Reset, accent, 50), accent);
-    }
-
-    #[test]
-    fn should_parse_path_short_flag() {
-        let parsed = parse_for_test(&["tuicr", "-p", "src/main.rs"]).expect("parse should succeed");
-        assert_eq!(parsed.path_filter, Some("src/main.rs".to_string()));
-    }
-
-    #[test]
-    fn should_parse_path_long_flag() {
-        let parsed = parse_for_test(&["tuicr", "--path", "src/"]).expect("parse should succeed");
-        assert_eq!(parsed.path_filter, Some("src/".to_string()));
-    }
-
-    #[test]
-    fn should_parse_path_equals_syntax() {
-        let parsed = parse_for_test(&["tuicr", "--path=plans/current-plan.md"])
-            .expect("parse should succeed");
-        assert_eq!(
-            parsed.path_filter,
-            Some("plans/current-plan.md".to_string())
-        );
-    }
-
-    #[test]
-    fn should_error_when_path_value_missing() {
-        let err = parse_for_test(&["tuicr", "--path"]).expect_err("parse should fail");
-        assert!(err.contains("--path requires a file or directory path"));
-    }
-
-    #[test]
-    fn should_error_when_path_equals_empty() {
-        let err = parse_for_test(&["tuicr", "--path="]).expect_err("parse should fail");
-        assert!(err.contains("--path requires a file or directory path"));
-    }
-
-    #[test]
-    fn should_default_path_filter_to_none() {
-        let parsed = parse_for_test(&["tuicr"]).expect("parse should succeed");
-        assert_eq!(parsed.path_filter, None);
-    }
-
-    #[test]
-    fn should_parse_path_with_working_tree() {
-        let parsed =
-            parse_for_test(&["tuicr", "-p", "file.md", "-w"]).expect("parse should succeed");
-        assert_eq!(parsed.path_filter, Some("file.md".to_string()));
-        assert!(parsed.working_tree);
-    }
-
-    #[test]
-    fn should_parse_path_with_revisions() {
-        let parsed = parse_for_test(&["tuicr", "--path", "src/", "-r", "HEAD~3.."])
-            .expect("parse should succeed");
-        assert_eq!(parsed.path_filter, Some("src/".to_string()));
-        assert_eq!(parsed.revisions, Some("HEAD~3..".to_string()));
-    }
-
-    #[test]
-    fn should_parse_pr_target_as_bare_number() {
-        // given/when
-        let parsed = parse_for_test(&["tuicr", "pr", "125"]).expect("parse should succeed");
-        // then
-        assert_eq!(parsed.pr_target, Some("125".to_string()));
-    }
-
-    #[test]
-    fn should_parse_pr_target_as_owner_repo_hash() {
-        // given/when
-        let parsed =
-            parse_for_test(&["tuicr", "pr", "agavra/tuicr#125"]).expect("parse should succeed");
-        // then
-        assert_eq!(parsed.pr_target, Some("agavra/tuicr#125".to_string()));
-    }
-
-    #[test]
-    fn should_parse_pr_target_as_full_url() {
-        // given/when
-        let parsed = parse_for_test(&["tuicr", "pr", "https://github.com/agavra/tuicr/pull/125"])
-            .expect("parse should succeed");
-        // then
-        assert_eq!(
-            parsed.pr_target,
-            Some("https://github.com/agavra/tuicr/pull/125".to_string()),
-        );
-    }
-
-    #[test]
-    fn should_error_when_pr_target_is_missing() {
-        // given/when
-        let err = parse_for_test(&["tuicr", "pr"]).expect_err("parse should fail");
-        // then
-        assert!(err.contains("tuicr pr requires a target"));
-    }
-
-    #[test]
-    fn should_error_when_pr_target_looks_like_flag() {
-        // given/when
-        let err = parse_for_test(&["tuicr", "pr", "--theme"]).expect_err("parse should fail");
-        // then
-        assert!(err.contains("tuicr pr requires a target"));
-    }
-
-    #[test]
-    fn should_combine_pr_target_with_theme_flag() {
-        // given/when — flag arguments still apply after the PR target.
-        let parsed = parse_for_test(&["tuicr", "pr", "125", "--theme", "dark"])
-            .expect("parse should succeed");
-        // then
-        assert_eq!(parsed.pr_target, Some("125".to_string()));
-        assert_eq!(parsed.theme, Some(ThemeArg::Dark));
-    }
-
-    #[test]
-    fn should_leave_pr_target_none_when_no_pr_subcommand() {
-        // given/when
-        let parsed = parse_for_test(&["tuicr"]).expect("parse should succeed");
-        // then
-        assert_eq!(parsed.pr_target, None);
-    }
-
-    #[test]
-    fn should_parse_repo_url_https() {
-        // given/when
-        let parsed = parse_for_test(&["tuicr", "--repo-url", "https://github.com/slatedb/slatedb"])
-            .expect("parse should succeed");
-        // then
-        assert_eq!(
-            parsed.repo_url,
-            Some("https://github.com/slatedb/slatedb".to_string())
-        );
-    }
-
-    #[test]
-    fn should_parse_repo_url_equals_form() {
-        // given/when
-        let parsed = parse_for_test(&["tuicr", "--repo-url=git@github.com:slatedb/slatedb.git"])
-            .expect("parse should succeed");
-        // then
-        assert_eq!(
-            parsed.repo_url,
-            Some("git@github.com:slatedb/slatedb.git".to_string())
-        );
-    }
-
-    #[test]
-    fn should_parse_repo_url_ssh_scheme() {
-        // given/when
-        let parsed = parse_for_test(&[
-            "tuicr",
-            "--repo-url",
-            "ssh://git@github.com/slatedb/slatedb.git",
-        ])
-        .expect("parse should succeed");
-        // then
-        assert_eq!(
-            parsed.repo_url,
-            Some("ssh://git@github.com/slatedb/slatedb.git".to_string())
-        );
-    }
-
-    #[test]
-    fn should_error_when_repo_url_value_missing() {
-        // given/when
-        let err = parse_for_test(&["tuicr", "--repo-url"]).expect_err("parse should fail");
-        // then
-        assert!(err.contains("--repo-url requires a value"));
-    }
-
-    #[test]
-    fn should_error_when_repo_url_value_is_flag() {
-        // given/when
-        let err =
-            parse_for_test(&["tuicr", "--repo-url", "--theme"]).expect_err("parse should fail");
-        // then
-        assert!(err.contains("--repo-url requires a value"));
-    }
-
-    #[test]
-    fn should_error_when_repo_url_unparseable() {
-        // given/when
-        let err =
-            parse_for_test(&["tuicr", "--repo-url", "not-a-url"]).expect_err("parse should fail");
-        // then
-        assert!(err.contains("not a recognized GitHub URL"));
-    }
-
-    #[test]
-    fn should_error_when_repo_url_equals_empty() {
-        // given/when
-        let err = parse_for_test(&["tuicr", "--repo-url="]).expect_err("parse should fail");
-        // then
-        assert!(err.contains("--repo-url requires a value"));
-    }
-
-    #[test]
-    fn should_leave_repo_url_none_when_not_provided() {
-        // given/when
-        let parsed = parse_for_test(&["tuicr"]).expect("parse should succeed");
-        // then
-        assert_eq!(parsed.repo_url, None);
     }
 }
